// src/toolbase-api-client/dto/operators.enum.ts
export enum EqualityOperator {
    EQUAL = '=',
    DOUBLE_EQUAL = '==',
    TRIPLE_EQUAL = '===',
    NOT_EQUAL = '<>',
    NOT_EQUAL_BANG = '!=',
    STRICT_NOT_EQUAL = '!=='
}

export enum NumericOperator {
    LESS_THAN = '<',
    LESS_THAN_EQUAL = '<=',
    GREATER_THAN = '>',
    GREATER_THAN_EQUAL = '>='
}